% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{.replaceNa}
\alias{.replaceNa}
\title{Replaces the NA values in a with corresponding values in b}
\usage{
.replaceNa(a, b)
}
\arguments{
\item{a, b}{objects to be tested or coerced.}
}
\value{
The altered object.
}
\description{
Replaces the NA values in a with corresponding values in b
}
