% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{.replaceNull}
\alias{.replaceNull}
\title{Uses b if a is NULL}
\usage{
.replaceNull(a, b)
}
\arguments{
\item{a, b}{objects to be tested or coerced.}
}
\value{
An un-null object.
}
\description{
Uses b if a is NULL
}
